added user

+++ src/main/java/com/nishchay/journalApp/repository/JournalEntryRepository.java
@@ -4,5 +4,5 @@ import com.nishchay.journalApp.entity.JournalEntry;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface JournalEntryRepository extends JpaRepository<JournalEntry, Integer> {
+public interface JournalEntryRepository extends JpaRepository<JournalEntry, Long> {
 }
